Revert "[@osd/cross-platform] Adds cross-platform helpers (#2681)"
This reverts commit 887093d2d243045029b644680a3e8d0150318143.

Now that downstream plugins and projects are unblocked from builds, we want to make these changes following our standard processes and automated checks
