+++ packages/osd-config-schema/package.json
@@ -10,9 +10,8 @@
     "osd:bootstrap": "yarn build"
   },
   "devDependencies": {
-    "@osd/cross-platform": "1.0.0",
-    "tsd": "^0.21.0",
-    "typescript": "4.0.2"
+    "typescript": "4.0.2",
+    "tsd": "^0.21.0"
   },
   "peerDependencies": {
     "lodash": "^4.17.21",
