--- packages/osd-cross-platform/src/index.ts
@@ -1,7 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './path';
-export * from './process';
