--- packages/osd-cross-platform/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "target",
-    "declaration": true,
-    "declarationMap": true
-  },
-  "include": [
-    "src/**/*"
-  ]
-}
