+++ packages/osd-plugin-generator/src/integration_tests/generate_plugin.test.ts
@@ -32,8 +32,7 @@ import Path from 'path';
 
 import del from 'del';
 import execa from 'execa';
-import { standardize } from '@osd/cross-platform';
-import { REPO_ROOT, createAbsolutePathSerializer } from '@osd/dev-utils';
+import { REPO_ROOT, standardize, createAbsolutePathSerializer } from '@osd/dev-utils';
 import globby from 'globby';
 
 // Has to be a posix reference because it is used to generate glob patterns
