+++ packages/osd-plugin-helpers/package.json
@@ -16,7 +16,6 @@
     "osd:watch": "tsc --watch"
   },
   "dependencies": {
-    "@osd/cross-platform": "1.0.0",
     "@osd/dev-utils": "1.0.0",
     "@osd/optimizer": "1.0.0",
     "del": "^5.1.0",
