+++ packages/osd-pm/src/utils/projects_tree.ts
@@ -31,7 +31,7 @@
 import chalk from 'chalk';
 import path from 'path';
 
-import { standardize } from '@osd/cross-platform';
+import { standardize } from '@osd/utils';
 import { Project } from './project';
 
 const projectKey = Symbol('__project');
